refactor: combine mobility and stations service
