+++ src/api/mobility/index.ts
@@ -1,7 +1,7 @@
 import Hapi from '@hapi/hapi';
 import { IMobilityService } from '../../service/interface';
-import { VehiclesQuery } from '../../service/types';
-import { getScootersRequest } from './schema';
+import { StationsQuery, VehiclesQuery } from '../../service/types';
+import { getScootersRequest, getStationsRequest } from './schema';
 
 export default (server: Hapi.Server) => (service: IMobilityService) => {
   server.route({
@@ -16,6 +16,20 @@ export default (server: Hapi.Server) => (service: IMobilityService) => {
       const payload = request.query as unknown as VehiclesQuery;
 
       return (await service.getVehicles(payload)).unwrap();
+    }
+  });
+  server.route({
+    method: 'GET',
+    path: '/bff/v2/mobility/stations',
+    options: {
+      tags: ['api', 'mobility', 'stations', 'bike', 'car'],
+      validate: getStationsRequest,
+      description: 'Get stations for bikes, car sharing etc.'
+    },
+    handler: async request => {
+      const payload = request.query as unknown as StationsQuery;
+
+      return (await service.getStations(payload)).unwrap();
     }
   });
 };
