--- src/api/stations/schema.ts
@@ -1,16 +1,0 @@
-import Joi from 'joi';
-import { FormFactor } from '../../graphql/mobility/mobility-types_v2';
-
-export const getStationsRequest = {
-  query: Joi.object({
-    availableFormFactors: Joi.array()
-      .items(Joi.string())
-      .optional()
-      .default(FormFactor.Bicycle)
-      .single(),
-    lat: Joi.number().required(),
-    lon: Joi.number().required(),
-    range: Joi.number().optional().default(500),
-    operators: Joi.array().items(Joi.string()).optional().single()
-  })
-};
